[#1005] Fixes for codacy
